Partie Frais Hors Forfait : transf tableau en formulaire + envoi corrigé et id FraisHorsForfait

+++ PROJETS PERSONNALISÉS ENCADRÉS (PPE) - SPÉCIALITÉ SLAM/www/vues/v_listeFraisHorsForfait.php
@@ -35,15 +35,23 @@
                 $montant = $unFraisHorsForfait['montant'];
                 $id = $unFraisHorsForfait['id']; ?>           
                 <tr>
+                <?php if(!$_SESSION['comptableBool']){?>
                     <td> <?php echo $date ?></td>
                     <td> <?php echo $libelle ?></td>
                     <td><?php echo $montant ?></td>
                     <td><a href="index.php?uc=gererFrais&action=supprimerFrais&idFrais=<?php echo $id ?>" 
                            onclick="return confirm('Voulez-vous vraiment supprimer ce frais?');">Supprimer ce frais</a></td>
+                <?php }else{?>
+                <form method="post" action="index.php?uc=validerFrais&action=corrigerFraisHorsForfait&idFraisHF=<?php echo $id ?>" role="form">
+                	<td><input type="text" id="dateHF" name="dateHF[<?php echo $id ?>]" value="<?php echo $date ?>"></td>
+                	<td><input type="text" id="libHF" name="libHF[<?php echo $id ?>]" value="<?php echo $libelle ?>"></td>
+                	<td><input type="text" id="montantHF" name="montantHF[<?php echo $id ?>]" value="<?php echo $montant ?>"></td>
+                	<td><button class="btn btn-success" type="submit"
+                 onclick="alert('Les informations de frais ont bien été corrigées');">Corriger</button></td>
+                 </form >
+                <?php } ?>
                 </tr>
-                <?php
-            }
-            ?>
+                <?php } ?>
             </tbody>  
         </table>
     </div>
